reestructuración de las carpetas del código fuente

+++ source/distribucionesDiscretas.cpp
@@ -24,23 +24,23 @@
 //#include "distribuciones.hpp"
 #include <math.h>
 
-double poisson(float parametroP, unsigned short int probabilidad) {
+double poisson(float parametroP, unsigned int probabilidad) {
 	double resultado;
 	float lambda = parametroP;
 	double e = 2.7182818284590;
-	unsigned short int x = probabilidad;
+	unsigned int x = probabilidad;
 	resultado = (pow(e, -lambda)) * (pow(lambda, x) / factorial(x));
 	return resultado;
 }
 
-double bernouilli(float p, unsigned short int x) {
+double bernouilli(float p, unsigned int x) {
 	double resultado;
 	resultado = potencia(p, x) * potencia(1 - p, x - 1);
 	return resultado;
 }
 
-double binomial(unsigned short int n, float p,
-                unsigned short int x) {
+double binomial(unsigned int n, float p,
+                unsigned int x) {
 	double resultado;
 	float q = (1 - p);
 	if(n > 50 && p <= 0.1) {
@@ -51,22 +51,22 @@ double binomial(unsigned short int n, float p,
 	return resultado;
 }
 
-double binomialNegativa(unsigned short int parametron, float parametroP,
-                        unsigned short int probabilidad) {
+double binomialNegativa(unsigned int parametron, float parametroP,
+                        unsigned int probabilidad) {
 	double resultado;
-	unsigned short int n = parametron;
+	unsigned int n = parametron;
 	float p = parametroP;
 	int x = probabilidad;
 	resultado = binomio(n + (x - 1), x) * potencia(p, n) * potencia(1 - p, x);
 	return resultado;
 }
 
-double hipergeometrica(unsigned short int parametroN, unsigned short int parametron,
-                       unsigned short int parametror,
-                       unsigned short int probabilidad) {
-	unsigned short int N = parametroN;
-	unsigned short int x = probabilidad;
-	unsigned short int n = parametron;
+double hipergeometrica(unsigned int parametroN, unsigned int parametron,
+                       unsigned int parametror,
+                       unsigned int probabilidad) {
+	unsigned int N = parametroN;
+	unsigned int x = probabilidad;
+	unsigned int n = parametron;
 	float r = parametror;
 	float parametroP = r/N;
 	double resultado;
@@ -78,11 +78,11 @@ double hipergeometrica(unsigned short int parametroN, unsigned short int paramet
 	return resultado;
 }
 
-double geometrica(float parametroP, unsigned short int probabilidad) {
+double geometrica(float parametroP, unsigned int probabilidad) {
 	double resultado;
 	float p = parametroP;
 	float q = 1 - p;
-	unsigned short int x = probabilidad;
+	unsigned int x = probabilidad;
 	resultado = (potencia(p * q, x - 1));
 	return resultado;
 }
